fix(dashboard): 인기가게, 매출 순위 구하는 로직 수정 (#65)

+++ src/main/java/com/example/wait4eat/domain/dashboard/entity/PopularStore.java
@@ -26,7 +26,7 @@ public class PopularStore {
     private String storeName;
 
     @Column(nullable = false)
-    private int waitingCount;
+    private Long waitingCount;
 
     @Column(nullable = false)
     private int ranking;
@@ -39,7 +39,7 @@ public class PopularStore {
     public PopularStore(
             Long storeId,
             String storeName,
-            int waitingCount,
+            Long waitingCount,
             int ranking,
             PopularityType type,
             Dashboard dashboard
